Revert "Breakup hooks into their own files (#211)"
This reverts commit 78ec0c3d63c1c529e7e690fd29ed7b40eef81d38.
