--- assets/js/hooks/alpine.js
@@ -1,14 +1,0 @@
-const Alpine = {
-  mounted() {
-    this.el.dispatchEvent(new CustomEvent('mounted', {
-      bubbles: true
-    }));
-  },
-  updated() {
-    this.el.dispatchEvent(new CustomEvent('updated', {
-      bubbles: true
-    }));
-  },
-};
-
-export default Alpine;
