--- assets/js/hooks/frame.js
@@ -1,17 +1,0 @@
-const Frame = {
-  mounted() {
-    const resizeObserver = new ResizeObserver(_entries => {
-      window.parent.postMessage({
-        height: this.el.scrollHeight
-      }, "*");
-    });
-
-    resizeObserver.observe(this.el);
-    window.parent.postMessage({
-      height: this.el.scrollHeight
-    }, "*");
-  }
-  
-}
-
-export default Frame;
