--- assets/js/hooks/index.js
@@ -1,30 +1,0 @@
-import TasksList from "./tasks_list";
-import RidersList from "./riders_list";
-import ConversationList from "./conversation_list";
-import MessageList from "./message_list";
-import RiderSelectionList from "./rider_selection_list";
-import Alpine from "./alpine";
-import Autocomplete from './autocomplete';
-import EmojiButton from './emoji_button';
-import CheckboxAll from './checkbox_all';
-import StatsDashboardChart from './stats_dashboard_chart';
-import Frame from './frame';
-import TagsComponent from './tags_component';
-import LeafletMap from './leaflet_map';
-
-
-export default {
-  TasksList,
-  RidersList,
-  ConversationList,
-  MessageList,
-  RiderSelectionList,
-  Alpine,
-  Autocomplete,
-  EmojiButton,
-  StatsDashboardChart,
-  CheckboxAll,
-  Frame,
-  TagsComponent,
-  LeafletMap
-}
