--- assets/js/hooks/tags_component.js
@@ -1,15 +1,0 @@
-const TagsComponent = {
-  mounted() {
-    this.el.addEventListener('keydown', e => {
-      if (e.key == 'Enter') {
-        e.preventDefault();
-        this.pushEventTo(this.el, 'select', {
-          name: this.el.value
-        })
-        this.el.value = ""
-      }
-    });
-  }
-};
-
-export default TagsComponent;
