--- assets/js/hooks/tasks_list.js
@@ -1,15 +1,0 @@
-const TasksList = {
-  mounted() {
-    this.handleEvent("select_task", ({
-      id
-    }) => {
-      this.el.querySelector(`[id="tasks-list:${id}"]`).scrollIntoView({
-        behavior: 'auto',
-        block: 'nearest',
-        inline: 'start'
-      })
-    })
-  }
-};
-
-export default TasksList;
