+++ lib/bike_brigade_web/live/calendar_live/index.html.heex
@@ -1,7 +1,7 @@
 <div
   id="calendar-frame"
   class="w-full p-2 text-base leading-7 text-white bg-black"
-  phx-hook="Frame"
+  phx-hook="FrameHook"
 >
   <div class="w-full py-3 text-center mb-14">
     <h1 class="font-serif text-4xl font-medium ">
