+++ lib/bike_brigade_web/live/sms_message_live/conversation_component.html.heex
@@ -180,7 +180,7 @@
               <div class="flex items-center justify-between space-x-2">
                 <.button
                   id="emoji-button"
-                  phx-hook="EmojiButton"
+                  phx-hook="EmojiButtonHook"
                   data-input-id={input_id(f, :body)}
                   type="button"
                   color={:white}
