+++ lib/bike_brigade_web/live/sms_message_live/form_component.html.heex
@@ -24,7 +24,7 @@
       <div class="flex justify-end mt-6 space-x-2">
         <.button
           id="emoji-button2"
-          phx-hook="EmojiButton"
+          phx-hook="EmojiButtonHook"
           data-input-id={input_id(f, :body)}
           type="button"
           color={:white}
