+++ lib/bike_brigade_web/live/stats_live/dashboard.html.heex
@@ -89,7 +89,7 @@
 <div class="h-full mt-4 bg-white rounded-lg shadow sm:p-6">
   <canvas
     id="rider-plot"
-    phx-hook="StatsDashboardChart"
+    phx-hook="Chart"
     phx-update="ignore"
     data-title="Stats"
     class="object-scale-down"
